refactor: controller에 커맨드 타입 도입
명령이 추가될 때마다 if문이 계속 증가하는 문제 발생
