+++ src/main/java/controller/AttendanceCommand.java
@@ -1,0 +1,7 @@
+package controller;
+
+import domain.Crews;
+
+public interface AttendanceCommand {
+    void execute(Crews crews);
+}
