+++ src/main/java/controller/command/AttendanceRecordByCrewCommand.java
@@ -1,0 +1,24 @@
+package controller.command;
+
+import controller.AttendanceCommand;
+import domain.Crew;
+import domain.Crews;
+import domain.Nickname;
+import view.InputView;
+import view.OutputView;
+
+public class AttendanceRecordByCrewCommand implements AttendanceCommand {
+
+    @Override
+    public void execute(final Crews crews) {
+        Nickname nickname = readNickname();
+        Crew crew = crews.findByNickname(nickname);
+
+        OutputView.printCrewAttendances(crew);
+    }
+
+    private Nickname readNickname() {
+        final String inputNickName = InputView.readNickName();
+        return new Nickname(inputNickName);
+    }
+}
